Create the /dev/trex character device.
This passes the currently failing bats test for the presence of
/dev/trex

This commit creates a character device in the device node tree. This
device currently does nothing. Actually, not strictly true, this device
currently crashes. Hard.

This commit does a couple of things. Firstly it defines some meta data
for the device that it's creating:

DEVICE_NAME: This will be the point in the /dev/ tree that gets created
DEVICE_CLASS: This creates a custom class for the device. I don't know
              exactly what this is yet, but from what I can see it's
              used to segregate different devices in sysfs. So with the
              custom class trex. I can see my device in /sys/class/trex/

Then it defines the interface for the device. The device is a character
device so we need to map some file operations for when the device is
opened for the first time, read from, written to and closed. This should
do us for now.

The device pointer is a struct that's defined within the kernel, and we
map the device instance the it's struct of file operations together
wihin the `register_chrdev` function call.

`register_chrdev` attempts to register a device with the kernel (and
assign a Major number for this driver, which is returned - these are the
numbers defined in /proc/devices, each major number represents a driver,
all devices with the same major number are controlled by the same
driver).

so the order of execution in the module init code is:

* register the driver (get a new Major number)
* create a char class
* create a new device node in that char class

Currently there is no code to tear all this down in the module exit
hook, so all this cruft will hang around until the machine is rebooted.

Also I have no idea how to go about finding an existing class and
registering a new device in that class, or even if that is a good idea.