Expand the tests
Now that the basic module load code is working, and creating a char
device this commit expands the test suite to define behavior for when
you read and write to the device.

Note that currently I need to reboot the VM every test run, because I
haven't written any module unload code!