Add instructions to access /dev/trex as non-root
We can't control this in the kernel code - it has to be handled by udev
when the kernel loads the driver.

So I've added a comment here that gives instructions on how to configure
udev to allow /dev/trex to be readable by all.

Dropping this config in place is probably an OS level packaging concern