Read the string "Rawr" from the device
This commit introduces the body of the read function. The behaviour we
want is to read the string "Rawr" from the kernel space to the user
space once and then stop.

because this is a char device and applications like `cat` will continue
to read out of the device until EOF is reached and all we want to do is
print out a string we have to track the number of times we've sent data
to the user and then fake an EOF.

EOF can be signalled by making the return value of the read function be
0.

This commit uses a `device_reads` integer for tracking how many times we
have called read from the device. It's initialised to 0 and then every
time we send a string to the user we increment it. If the count is
1 (ie. the read function has been called once since the device was
opened) then we know we've already printed the string so we can return 0
for EOF